# backend/openapi.yaml
openapi: "3.1.0"
info:
  title: Syspro APS Scheduler API
  version: "1.0.0"
  description: |
    Finite-capacity production scheduler that reads master data from SYSPRO
    and writes approved schedules back via the APS compatibility layer.
  contact:
    name: APS Development Team
servers:
  - url: http://localhost:3000/api
    description: Local development / production (same port)
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, format: password }
    LoginResponse:
      type: object
      properties:
        accessToken:  { type: string }
        refreshToken: { type: string }
        user:
          type: object
          properties:
            sub:      { type: integer }
            username: { type: string }
            role:     { type: string, enum: [Reviewer, Approver, super_admin] }

    # ── Jobs ──────────────────────────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        error:   { type: string }
        code:    { type: string }
        details: { type: object }
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
paths:
...
  /auth/login:
    post:
      summary: Obtain JWT access + refresh tokens
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        "200":
          description: Tokens issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        "401":
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      summary: Exchange refresh token for new access token
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken: { type: string }
        "401":
          $ref: '#/components/responses/Unauthorized'

  # Health
...